Replaced hacky C++03 lambdas with C++11 lambdas

+++ Client/core/CMainMenu.cpp
@@ -1121,7 +1121,7 @@ void CMainMenu::WantsToDisconnectCallBack( void* pData, uint uiButton )
 
     if ( uiButton == 1 )
     {
-        uchar menuType = (uchar)pData;
+        int menuType = (int)pData;
         switch( menuType )
         {
             case MENU_ITEM_HOST_GAME:       OnHostGameButtonClick ();    break;
